rename: move RoleModal, RoleSelection component
